# kratos/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Api API
    version: 0.0.1
paths:
    /api/create_account:
        post:
            tags:
                - Api
            operationId: Api_CreateAccount
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.CreateAccountRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.CreateAccountResponse'
    /health:
        get:
            tags:
                - Api
            operationId: Api_Health
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.HealthResponse'
    /info:
        get:
            tags:
                - Api
            operationId: Api_Info
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.InfoResponse'
    /login:
        post:
            tags:
                - Api
            operationId: Api_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.LoginResponse'
components:
    schemas:
        api.CreateAccountRequest:
            type: object
            properties:
                ownerId:
                    type: string
                balance:
                    type: number
                    format: double
        api.CreateAccountResponse:
            type: object
            properties:
                id:
                    type: string
                ownerId:
                    type: string
                balance:
                    type: number
                    format: double
        api.HealthResponse:
            type: object
            properties:
                status:
                    type: string
        api.InfoResponse:
            type: object
            properties:
                appName:
                    type: string
        api.LoginRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
        api.LoginResponse:
            type: object
            properties:
                username:
                    type: string
                role:
                    type: string
                resources:
                    type: array
                    items:
                        type: string
                token:
                    type: string
tags:
    - name: Api